requirements.txt
appnope==0.1.3
asgiref==3.5.2
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
blurhash-python==1.2.0
boto3==1.26.156
botocore==1.29.156
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.1
django-ckeditor==6.6.1
django-cors-headers==3.13.0
django-filter==22.1
django-gcloud-storage==0.5.0
django-js-asset==2.1.0
django-money==3.0.0
django-oauth-toolkit==2.2.0
django-parler==2.3
django-parler-rest==2.2
django-random-id-model==0.1.1
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
drf-social-oauth2==2.1.2
drf-yasg==1.21.4
ecdsa==0.18.0
executing==1.2.0
filelock==3.8.0
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-core==2.3.0
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
nanoid==2.0.0
networkx==2.8.8
numpy==1.24.3
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipenv==2022.10.4
platformdirs==2.5.2
prompt-toolkit==3.0.33
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-moneyed==2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydig==0.4.0
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.0
python-dotenv==0.12.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.3.2.post1
sqlparse==0.4.2
static3==0.7.0
stripe==5.0.0
toml==0.10.2
traitlets==5.5.0
typing_extensions==4.4.0
Unipath==1.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.15.0
xmltodict==0.13.0
